% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.r
\name{compile.lgcpSPDE}
\alias{compile.lgcpSPDE}
\title{Compiling TMB C++ templates}
\usage{
compile.lgcpSPDE()
}
\description{
Compiles the \code{lgcpSPDE} TMB templates into a shared object
file. This must be done a single time following installation or
updating of the package.
}

