% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{earthquakes}
\alias{earthquakes}
\title{A dataset taken from the GeoNet Quake (http://quakesearch.geonet.org.nz/)
containing earthquake information in New Zealand (and the Chatham islands) from 26-March-2018 to 26-April-2018}
\format{A data frame containing 368 rows and 5 variables:
\describe{
\item{latitude}{latitude location}
\item{longitude}{longitude location}
\item{origintime}{The UTC time of the event's occurrence}
\item{depth}{The focal depth of the event (km)}
\item{magnitude}{The magnitude of the earthquake}
}}
\usage{
data(earthquakes)
}
\description{
A dataset taken from the GeoNet Quake (http://quakesearch.geonet.org.nz/)
containing earthquake information in New Zealand (and the Chatham islands) from 26-March-2018 to 26-April-2018
}
