% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{find.fields}
\alias{find.fields}
\title{Function that extracts the ''random fields'' of the model fitted.}
\usage{
find.fields(x = NULL, mesh = NULL, n.t = NULL, sd = FALSE,
  plot = FALSE, spatial.polygon = NULL, ...)
}
\arguments{
\item{x}{A fitted model from \link{mark.pp.fit}().}

\item{mesh}{the mesh used in the model fit.}

\item{n.t}{numeric, the number of time points.}

\item{sd}{Logical, if \code{FALSE} means of random fields aer returned.}

\item{plot}{Logical, if \code{TRUE} the returned matricies (either SD or Mean of
random fields are plotted.}

\item{spatial.polygon}{Optional, if a spatial polygon of the domain is supplied, only
values of the random field within the domain will be returned}

\item{...}{additional graphical parameters
@importFrom fields image.plot}
}
\description{
Plots the estimated random fields or parameter posterior densities from an object returned by
\link{mark.pp.fit}().
}

