% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lgcp.r
\name{fit.lgcp}
\alias{fit.lgcp}
\title{Function to fit a  spatio-temporal log-Gaussian Cox process model  with an intercept and covariates (optional)}
\usage{
fit.lgcp(mesh = NULL, mesh.pars = NULL, locs = NULL, temp = NULL,
  covariates = NULL, prior.rho = list(theta = list(prior = "pccor1", param =
  c(0, 0.9))), verbose = FALSE, control.inla = list(strategy = "gaussian",
  int.strategy = "eb"), return.attributes = FALSE, ns = NULL)
}
\arguments{
\item{mesh}{a ``mesh'' object i.e. delauney triangulation of the domain, an
object returned by \link{make.mesh}.}

\item{mesh.pars}{a named vertor of mesh parameters, must contain
\code{cutoff} length at which to cut off triangle edge lengths,
\code{min} triangle edge length inside region,
and \code{max} triangle edge length outside region.}

\item{locs}{a matrix of observation locations, where each row corresponds to the observation.}

\item{temp}{a numeric vector specifying a temporal index for each observation (starting at 1.....T).}

\item{covariates}{a named data.frame of covariates}

\item{prior.rho}{prior for the temporal correlation coefficient, by default a \code{pcprior} is used with \code{param=c(0-0.9)}.}

\item{verbose}{Logical if \code{TRUE} model fit is output to screen.}

\item{response}{a vector of response variable, each corresponds to the spatial locations
in \code{locs}.}
}
\value{
A \code{inla} result object
}
\description{
Function to fit a  spatio-temporal log-Gaussian Cox process model  with an intercept and covariates (optional)
}
