% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_pp_fit.r
\name{fit.marked.lgcp}
\alias{fit.marked.lgcp}
\title{Fitting a marked point process model}
\usage{
fit.marked.lgcp(mesh = NULL, locs = NULL, t.index = NULL, mark = NULL,
  covariates = NULL, mark.family = "gaussian", verbose = FALSE,
  prior.rho = list(theta = list(prior = "pccor1", param = c(0, 0.9))),
  hyper = list(theta = list(prior = "normal", param = c(0, 10))),
  control.compute = list(dic = TRUE, waic = TRUE, cpo = TRUE, config = TRUE),
  control.inla = list(strategy = "gaussian", int.strategy = "eb"),
  link = NULL, ...)
}
\arguments{
\item{mesh}{delauney triangulation of area, an object returned by \link{make.mesh} is suitable.}

\item{locs}{a matrix of \code{nrow} locations in \code{ncol} dimesions.}

\item{t.index}{a vector of length \code{nrow} of time index units refering to each point location
given in \link{locs}.}

\item{mark}{a vector of length \code{nrow} of marks refering to each point location}

\item{mark.family}{assumed likelihood for mark, by defalt "gaussian".
#' @param hyper prior for the copy parameter by default is a N(0,10) i.e.,  list(theta=list(prior='normal', param=c(0,10)))}

\item{verbose}{Logical, if \code{TRUE}, model fitting is output
the console.}

\item{control.compute}{a list of fit statistics the user wants INLA to return. By default this
is \code{list(dic = TRUE, waic = TRUE,cpo = TRUE, config = TRUE)}.}

\item{control.inla}{a list which controls the fitting procedures INLA uses see Rue et al. ***ref book***
by default this is \code{list(strategy='gaussian',int.strategy = 'eb')} for quick and dirty fitting.}
}
\value{
An inla model fit object,
}
\description{
Fits a spatio-temporal marked point process model
using INLA coupled with the SPDE approch
}

