% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_joint_fit.r
\name{geo.joint.fit}
\alias{geo.joint.fit}
\title{Function to fit a joint spatio-temporal model to geo-statistical data where one spatio-
temporal component is shared between the responses}
\usage{
geo.joint.fit(mesh = NULL, locs.1 = NULL, locs.2 = NULL,
  response.1 = NULL, response.2 = NULL, t.index = NULL,
  covariates = NULL, family = c("gaussian", "gaussian"), verbose = FALSE,
  prior.rho = list(theta = list(prior = "pccor1", param = c(0, 0.9))),
  hyper = list(theta = list(prior = "normal", param = c(0, 10))),
  control.inla = list(strategy = "gaussian", int.strategy = "eb"))
}
\arguments{
\item{mesh}{a ``mesh'' object i.e. delauney triangulation of the domain, an
object returned by \link{make.mesh}.}

\item{covariates}{a named data.frame of covariates}

\item{family}{a character vector specifying the assumed likelihoods of the response, by default is c("gaussian,"gaussian").}

\item{verbose}{Logical if \code{TRUE} model fit is output to screen.}

\item{prior.rho}{prior for the temporal correlation coefficient, by default a \code{pcprior} is used with \code{param=c(0-0.9)}.}

\item{locs}{a matrix of observation locations, where each row corresponds to the observation.}

\item{response}{a matrix of response variables, each row corresponds to the spatial locations
in \code{locs}, the first column is the first response as a result of a
spatio-temporal process we want to ``copy'' into the second column response.}

\item{temp}{a numeric vector specifying a temporal index for each observation (starting at 1.....T)}
}
\value{
A \code{inla} result object
}
\description{
Function to fit a joint spatio-temporal model to geo-statistical data where one spatio-
temporal component is shared between the responses
}

