% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_fit.r
\name{geo.spatial.fit}
\alias{geo.spatial.fit}
\title{spatial only model fitting}
\usage{
geo.spatial.fit(mesh, locs, response, covariates, family, control.inla,
  control.compute, non.linear, prediction, sig0, Psig, rho0, Prho, verbose, ...)
}
\description{
spatial only model fitting
}

