% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_joint_fit.r
\name{geo.spatial.j.fit}
\alias{geo.spatial.j.fit}
\title{spatial only fitting}
\usage{
geo.spatial.j.fit(mesh, locs, response, covariates, family, control.inla, hyper,
  control.compute, sig0, rho0, verbose)
}
\description{
spatial only fitting
}

