% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_joint_fit.r
\name{geo.spatial.j.temporal.fit}
\alias{geo.spatial.j.temporal.fit}
\title{spatio-temporal model fitting}
\usage{
geo.spatial.j.temporal.fit(mesh, locs, response, covariates, temp, family,
  control.time, control.inla, hyper, control.compute, sig0, rho0, verbose)
}
\description{
spatio-temporal model fitting
}

