% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_fit.r
\name{geo.spatial.temporal.fit}
\alias{geo.spatial.temporal.fit}
\title{spatio temporal model fitting}
\usage{
geo.spatial.temporal.fit(mesh, locs, response, covariates, temp, family,
  control.time, control.inla, control.compute, non.linear, prediction, sig0,
  Psig, rho0, Prho, verbose, ...)
}
\description{
spatio temporal model fitting
}

