% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_st_fit.r
\name{geo.st.fit}
\alias{geo.st.fit}
\title{Function to fit a  spatio-temporal model to geo-statistical data with an intercept and covariates}
\usage{
geo.st.fit(mesh = NULL, locs = NULL, response = NULL, temp = NULL,
  covariates = NULL, family = "gaussian", prior.rho = list(theta =
  list(prior = "pccor1", param = c(0, 0.9))), verbose = FALSE)
}
\arguments{
\item{mesh}{a ``mesh'' object i.e. delauney triangulation of the domain, an
object returned by \link{make.mesh}.}

\item{locs}{a matrix of observation locations, where each row corresponds to the observation.}

\item{response}{a vector of response variable, each corresponds to the spatial locations
in \code{locs}.}

\item{temp}{a numeric vector specifying a temporal index for each observation (starting at 1.....T).}

\item{covariates}{a named data.frame of covariates}

\item{family}{a character vector specifying the assumed likelihood of the response, by default is "gaussian".}

\item{prior.rho}{prior for the temporal correlation coefficient, by default a \code{pcprior} is used with \code{param=c(0-0.9)}.}

\item{verbose}{Logical if \code{TRUE} model fit is output to screen.}
}
\value{
A \code{inla} result object
}
\description{
Function to fit a  spatio-temporal model to geo-statistical data with an intercept and covariates
}

