% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{lgcp2D}
\alias{lgcp2D}
\title{A simulated realisation of a log-Gaussian Cox process 2D}
\format{A data frame containing 281 rowa and 2 variables:
\describe{
\item{x}{x-coordinate of the point pattern}
\item{y}{y-coordinate of the point pattern}
}}
\usage{
data(lgcp2D)
}
\description{
A simulated realisation of a log-Gaussian Cox process 2D
}

