% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mesh.r
\name{make.mesh}
\alias{make.mesh}
\title{Wrapper function to create a mesh object as used in an INLA model fit}
\usage{
make.mesh(locs = NULL, mesh.pars = NULL, spatial.polygon = NULL,
  sphere = FALSE, plot = FALSE)
}
\arguments{
\item{locs}{a matrix of locations, either this or \code{spatial.polygon} must be supplied.}

\item{mesh.pars}{a named vertor of mesh parameters, must contain
\code{cutoff} length at which to cut off triangle edge lengths,
\code{min} triangle edge length inside region,
and \code{max} triangle edge length inside region.}

\item{spatial.polygon}{if supplied the spatial polygon for the domain is used to construct mesh}

\item{sphere}{Logical if TRUE the mesh is constructed on the unit sphere, note this is only possible if coordinates are
longitude and Latitude, by default FALSE}

\item{plot}{Logical if TRUE the triangulation is plotted}
}
\value{
A ``mesh'' object used in an INLA/SPDE model
}
\description{
Wrapper function to create a mesh object as used in an INLA model fit
}

