% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{NZ}
\alias{NZ}
\title{Spatial polygon of New Zealand (including the Chatham islands)}
\format{A Spatial Polygons Data Frame of New Zealand}
\usage{
data(NZ)
}
\description{
Spatial polygon of New Zealand (including the Chatham islands)
}
\keyword{datasets}

