% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plot.fields}
\alias{plot.fields}
\title{Plots fields eith spatio-temporal or spatial}
\usage{
plot.fields(x = NULL, mesh = NULL, n.t = NULL, sd = FALSE,
  spatial.polygon = NULL, col = grey.colors(100, 0.05, 0.95), ...)
}
\arguments{
\item{x}{A fitted model from \link{mark.pp.fit}().}

\item{mesh}{the mesh used in the model fit.}

\item{n.t}{numeric, the number of time points.}

\item{sd}{Logical, if \code{FALSE} means of random fields aer returned.}

\item{spatial.polygon}{Optional, if a spatial polygon of the domain is supplied, only
values of the random field within the domain will be returned}

\item{...}{additional graphical parameters
@importFrom fields image.plot}
}
\description{
Plots fields eith spatio-temporal or spatial
}

