% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlgcpspde.r
\name{rlgcpspde}
\alias{rlgcpspde}
\title{Function to simulate from a spatial/spatio temporal LGCP (SPDE model)}
\usage{
rlgcpspde(spatial.polygon = NULL, mesh.pars = NULL, mu = 0,
  kappa = NULL, sigma2 = 0.05, n = 1, rho = 0.9, mark = FALSE,
  beta = NULL, mark.function = function(x, y) cos(x) - sin(y), seed = 1,
  non.stat = NULL)
}
\arguments{
\item{spatial.polygon}{the spatial polygon for the domain is used to construct the delauney traingulation}

\item{mesh.pars}{a named vertor of mesh parameters, must contain
\code{cutoff} length at which to cut off triangle edge lengths,
\code{min} triangle edge length inside region,
and \code{max} triangle edge length inside region.}

\item{mu}{numeric or a named list, the intercept term to simulate a LGCP, by default is 0.
if a named list must contain elements \code{mean}, \code{cov.model}, \code{cov.pars} the latter
 two parameters of the grf function \link{geoR}. to simulate a non-stationary expectation}

\item{kappa}{a numeric constant, parameter of the SPDE model.}

\item{sigma2}{a numeric constant, parameter of the SPDE model, by default this is 0.05.}

\item{n}{a numeric constant defining the number of time points, by default 1.}

\item{rho}{the ar1 correlation coefficient for spatio-temporal samples,
by default this is 0.9.}

\item{mark}{Logical, if TRUE a marked point pattern is simulated}

\item{beta}{a scalar, this, the interaction parameter describing the dependance between the mark and point locations}

\item{mark.function}{a function of 2D spatial coordinates which describes the spatial process
specific to the mark, by default this is \code{function(x,y) cos(x) - sin(y)}.}

\item{seed}{seed for the simulation, by default this is 1}

\item{non.stat}{a named list the first element \code{fn} the spatial function which kappa varies with
and \code{theta} a vctor of length 3 specifying the theta values of the non-stationary model}
}
\value{
A named matrix (or a list of matricies if spatio-temporal) of point locations and
(if a marked point pattern is simulated) a mark values
}
\description{
Function to simulate from a spatial/spatio temporal LGCP (SPDE model)
}

