% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{terrorism_aggregate}
\alias{terrorism_aggregate}
\title{A cleaned aggregated dataset orginally taken from the global terrorism database (GTD)
(http://www.start.umd.edu/gtd/)
containing information of terrorism activity worldwide 2010--2017}
\format{A data frame containing 10845 rows and 20 variables:
\describe{
\item{latitude}{latitude location}
\item{longitude}{longitude location}
\item{iyear}{year index 2010--2016}
\item{total}{total number of events}
\item{lethal}{number of lethal events}
\item{imonth}{numeric month index 1--12}
\item{iday}{numeric day 1--31 (zeros are a non-entry)}
\item{country}{country}
\item{sucess}{Logical fatal or not. TRUE = fatal}
\item{nkill}{total number of fatalities}
\item{specificity}{factor which represents the apatial accuracy of the evets: 1 = most accurate, 5 = worst}
\item{gname}{character name of attack perpetrators}
\item{x.coord}{x coordinate from location projected onto a sphere}
\item{y.coord}{y coordinate from location projected onto a sphere}
\item{z.coord}{z coordinate from location projected onto a sphere}
\item{popdensity}{scaled: number of people per kilometer squared}
\item{luminosity}{scaled: luminosity}
\item{tt}{scaled: time to nearest city in minutes}
\item{country.idx}{numeric country index}
\item{originalcount}{total number of fatalities}
}}
\usage{
data(terrorism_aggregate)
}
\description{
A cleaned aggregated dataset orginally taken from the global terrorism database (GTD)
(http://www.start.umd.edu/gtd/)
containing information of terrorism activity worldwide 2010--2017
}
