% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{terrorism_data}
\alias{terrorism_data}
\title{A cleaned aggregated dataset orginally taken from the global terrorism database (GTD)
(http://www.start.umd.edu/gtd/)
containing information of terrorism activity worldwide 2010--2016}
\format{A data frame containing 8647 rows and 14 variables:
\describe{
\item{latitude}{latitude location}
\item{longitude}{longitude location}
\item{iyear}{year index 2010--2016}
\item{total}{total number of events}
\item{lethal}{number of lethal events}
\item{nkill}{total number of fatalities}
\item{alt}{altitude}
\item{pop}{population}
\item{tt}{time to nearest city (hours)}
\item{lum}{luminosity}
\item{country}{country code}
\item{x.coord}{x coordinate of location projected onto a sphere}
\item{y.coord}{y coordinate of location projected onto a sphere}
\item{z.coord}{z coordinate of location projected onto a sphere}
}}
\usage{
data(terrorism_data)
}
\description{
A cleaned aggregated dataset orginally taken from the global terrorism database (GTD)
(http://www.start.umd.edu/gtd/)
containing information of terrorism activity worldwide 2010--2016
}
