% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{world}
\alias{world}
\title{Spatial polygon of countries worldwide}
\format{A Spatial Polygons Data Frame of contries worldwide}
\usage{
data(world)
}
\description{
Spatial polygon of countries worldwide
}
\keyword{datasets}

