% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_pp_fit.r
\name{fit.marked.lgcp}
\alias{fit.marked.lgcp}
\title{Fitting a marked point process model}
\usage{
fit.marked.lgcp(mesh = NULL, locs = NULL, t.index = NULL, mark = NULL,
  covariates = NULL, mark.family = "gaussian", verbose = FALSE,
  prior.rho = list(theta = list(prior = "pccor1", param = c(0, 0.9))),
  prior.range = c(400, 0.5), prior.sigma = c(1, 0.05), hyper = list(theta
  = list(prior = "normal", param = c(0, 10))), control.compute = list(dic =
  TRUE, waic = TRUE, cpo = TRUE, config = TRUE), control.inla = list(strategy
  = "gaussian", int.strategy = "eb"), link = NULL, pp.int = FALSE,
  mark.int = FALSE, ...)
}
\arguments{
\item{mesh}{delauney triangulation of area, an object returned by \link{make.mesh} is suitable.}

\item{locs}{a matrix of \code{nrow} locations in \code{ncol} dimesions.}

\item{t.index}{a vector of length \code{nrow} of time index units refering to each point location
given in \link{locs}.}

\item{mark}{a vector of length \code{nrow} of marks refering to each point location}

\item{covariates}{a matrix of covariates (for the mark) with named columns.}

\item{mark.family}{assumed likelihood for mark, by defalt "gaussian". Ignored if \code{mark} is NULL.}

\item{verbose}{Logical, if \code{TRUE}, model fitting is output
the console.}

\item{prior.range}{pc prior for the range of the latent field (rnage0,Prange) (i.e., P(range < range0) = Prange
NOTE should be changed to reflect range of the domain by default this is 400, see Martins, Thiago G., et al.
"Penalising model component complexity: A principled, practical approach to constructing priors." (2014) for more details.}

\item{prior.sigma}{pc prior for the sd of the latent field (sigma0,Psigma) by default c(1,0.05) i.e., prob sigma > 1 = 0.05}

\item{hyper}{prior for the copy parameter by default is a N(0,10) i.e.,  list(theta=list(prior='normal', param=c(0,10)))}

\item{control.compute}{a list of fit statistics the user wants INLA to return. By default this
is \code{list(dic = TRUE, waic = TRUE,cpo = TRUE, config = TRUE)}.}

\item{control.inla}{a list which controls the fitting procedures INLA uses see Blangiardo, Marta, et al.
"Spatial and spatio-temporal models with R-INLA." Spatial and spatio-temporal epidemiology 4 (2013)
by default this is \code{list(strategy='gaussian',int.strategy = 'eb')} for quick and dirty fitting.}

\item{pp.int}{Logical, should an intercept term for the lgcp be included. By default FALSE. If TURE alpha0 estimated.}

\item{mark.int}{Logical, should an interpect term for the mark be included. By default FALSE. If TRUE beta0 estimated.}

\item{...}{other \code{inla} arguments}
}
\value{
An inla model fit object,
}
\description{
Fits a spatio-temporal marked point process model
using INLA coupled with the SPDE approch
}
