% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_pp_fit.r
\name{fit.marked.lgcp}
\alias{fit.marked.lgcp}
\title{Fitting a marked point process model}
\usage{
fit.marked.lgcp(mesh = NULL, locs = NULL, t.index = NULL, mark = NULL,
  covariates = NULL, mark.family = "gaussian", verbose = FALSE,
  prior.rho = list(theta = list(prior = "pccor1", param = c(0, 0.9))),
  hyper = list(theta = list(prior = "normal", param = c(0, 10))))
}
\arguments{
\item{mesh}{delauney triangulation of area, an object returned by \link{make.mesh} is suitable.}

\item{locs}{a matrix of \code{nrow} locations in \code{ncol} dimesions.}

\item{t.index}{a vector of length \code{nrow} of time index units refering to each point location
given in \link{locs}.}

\item{mark}{a vector of length \code{nrow} of marks refering to each point location}

\item{mark.family}{assumed likelihood for mark, by defalt "gaussian".}

\item{verbose}{Logical, if \code{TRUE}, model fitting is output
the console.}
}
\value{
An inla model fit object,
}
\description{
Fits a spatio-temporal marked point process model
using INLA coupled with the SPDE approch
}

